Configure ruby-lsp
ruby-lsp from shopify[1] seems to work well for what I was using
rbtagger TAGS files for - namely jumping to the definition of a
method.

I find the pop-up eldoc buffer distracting, so prefer to open it with
M-x eldoc for now.

I've also removed cape-keyword from the corfu completions as eglot
also provides that in a smarter way.

[1] https://github.com/Shopify/ruby-lsp